+++ frontend/app/src/main/java/com/example/myapplication/dto/trip/TripSuggestInfosDto.kt
@@ -3,7 +3,7 @@ package com.example.myapplication.dto.trip
 data class TripSuggestInfosDto(
     val id: Int?,
     val steps: List<TripStepDetailDto>,
-    val totalDuration: Int,
+    val totalDuration: Double,
     val totalCost: Int,
     val totalStep: Int,
     val weather: String,
